refactor(BikeGoal): Extract code in BikeGoalViewDesktop component
no code change
