+++ src/components/Views/BikeGoal.jsx
@@ -1,9 +1,7 @@
 import React from 'react'
-import { useNavigate } from 'react-router-dom'
 
 import { hasQueryBeenLoaded, isQueryLoading, useQuery } from 'cozy-client'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
-import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 
 import {
   // buildOneYearBikeCommuteTimeseriesQueryByDateAndAccountId,
@@ -11,16 +9,11 @@ import {
 } from 'src/queries/queries'
 import { useAccountContext } from 'src/components/Providers/AccountProvider'
 import SpinnerOrEmptyContent from 'src/components/SpinnerOrEmptyContent'
-import Titlebar from 'src/components/Titlebar'
-import BikeGoalList from 'src/components/Goals/BikeGoal/BikeGoalList'
 import BikeGoalDialogMobile from 'src/components/Goals/BikeGoal/BikeGoalDialogMobile'
-import BikeGoalAchievement from 'src/components/Goals/BikeGoal/BikeGoalAchievement'
-import BikeGoalChart from 'src/components/Goals/BikeGoal/BikeGoalChart'
+import BikeGoalViewDesktop from 'src/components/Goals/BikeGoal/BikeGoalViewDesktop'
 
 const Bikegoal = () => {
   const { account, isAccountLoading } = useAccountContext()
-  const { t } = useI18n()
-  const navigate = useNavigate()
   const { isMobile } = useBreakpoints()
 
   // TODO: uncomment this when the request return something
@@ -75,23 +68,10 @@ const Bikegoal = () => {
   }
 
   return (
-    <>
-      <Titlebar
-        label={t('bikeGoal.title')}
-        subtitle={<BikeGoalAchievement timeseries={timeseries} />}
-        onBack={() => navigate('/')}
-      />
-      <BikeGoalChart
-        className="u-flex u-flex-justify-end u-pos-absolute u-top-xl u-right-xl"
-        timeseries={timeseries}
-      />
-      <BikeGoalList
-        className="u-mt-3"
-        timeseries={timeseries}
-        hasMore={timeseriesQueryLeft.hasMore}
-        fetchMore={timeseriesQueryLeft.fetchMore}
-      />
-    </>
+    <BikeGoalViewDesktop
+      timeseries={timeseries}
+      timeseriesQueryLeft={timeseriesQueryLeft}
+    />
   )
 }
